Add internal drawer watch, and stop listening for events at the component level
We're listening for drawer toggle events at the `App.vue` level so there's no need to listen and respond to them at the `NavigationDrawer.vue` level.

Instead, we watch for when the variable is changed internally within the component, and then message that change as a global event to keep state values consistent.